* api refactor + remove duplicates (#140)

Enhancements on WebUI, API, Core, Modules, Lint, and Bugs fixed.

* Bump codecov from 2.1.8 to 2.1.9 (#155)

+++ core/load.py
@@ -855,12 +855,14 @@ def load_honeypot_engine():
         run_as_test
     )
 
+    # killed the network traffic capture process by ctrl + c... waiting to end.
+    info("killing network capture process")
+    if run_as_test:
+        network_traffic_capture_process.terminate()
+    # without ci it will be terminate after a few seconds, it needs to kill the tshark and update pcap file collection
+    network_traffic_capture_process.join()
     # if in case any events that were not inserted from thread
     push_events_queues_to_database(honeypot_events_queue, network_events_queue)
-    # kill the network traffic capture process
-    network_traffic_capture_process.terminate()
-    network_traffic_capture_process.join()
-    info("killing network capture process")
     # Kill bulk events thread
     terminate_thread(bulk_events_thread)
     # stop created containers
